backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-timezone-field==6.0.1
pytz==2023.3
requests==2.31.0